pyproject.toml
[project]
name = "vogonpoetry"
version = "0.1.0"
description = "YAML pipelines for sarcastic automation"
readme = "README.md"
keywords = ["llm", "ai", "mcp", "automation"]
requires-python = ">= 3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "pydantic>=2.11.7",
    "pymetrics>=1.0.10",
    "structlog (>=25.4.0,<26.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]
license = "MIT"
authors = [
    {name = "Teagan", email = "that@teagantotally.rocks"},
]
maintainers = [
    {name = "Teagan", email = "that@teagantotally.rocks"},
]

[tool.poetry]
name = "vogonpoetry"
version = "0.1.0"
description = "YAML pipelines for sarcastic automation"
authors = ["Teagan <teagan@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
pydantic = "^2.11.7"
pymetrics = "^1.0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[tool.poetry.group.prometheus.dependencies]
prometheus_client = "^0.18.0"

[tool.poetry.group.sentencetransformers.dependencies]
sentence_transformers = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"